+++ forge/src/main/java/whocraft/tardis_refined/common/data/LangProviderEnglish.java
@@ -291,6 +291,7 @@ public class LangProviderEnglish extends LanguageProvider {
         addShell(ShellTheme.HIEROGLYPH.getId(), "Hieroglyph");
         addShell(ShellTheme.CASTLE.getId(), "Castle");
         addShell(ShellTheme.PATHFINDER.getId(), "Pathfinder");
+        addShell(ShellTheme.HALF_BAKED.getId(), "Half Baked");
 
         /*Tool Tips*/
         add(ModMessages.TOOLTIP_TARDIS_LIST_TITLE, ChatFormatting.YELLOW + "< "